requirements.txt
aiogram==3.22.0
aiohttp==3.9.1
sqlalchemy==2.0.43
asyncpg==0.30.0
python-dotenv==1.0.0
httpx==0.26.0
apscheduler==3.11.0